move env.ts file into config folder
